Add opaque types for delta time, viewport units, GL locations; quality sweep
Seconds opaque type for delta time across all traits and implementations.
WorldUnits opaque type for viewport dimensions. UniformLocation and
AttributeLocation opaque types for GL shader variable type safety.
Remove deprecated methods, redundant setColor overloads. Convert
singletons to objects.

+++ sge/src/main/scala/sge/graphics/g3d/particles/ParticleShader.scala
@@ -447,7 +447,7 @@ object ParticleShader {
       override def set(shader: BaseShader, inputID: Int, renderable: Renderable, combinedAttributes: Attributes): Unit =
         // Use camera viewport width as approximation for screen width
         shader.camera.foreach { cam =>
-          shader.setFloat(inputID, cam.viewportWidth)
+          shader.setFloat(inputID, cam.viewportWidth.toFloat)
         }
     }
